Move diff, digitize, isin
